reflection-agent/pyproject.toml
[tool.poetry]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = ["adamscarlat <ascarlat@vividcloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
langchain = "^0.2.7"
langchain-openai = "^0.1.14"
langgraph = "^0.1.5"
grandalf = "^0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
